+++ R/plot.R
@@ -1,3 +1,163 @@
+#' Plot array live times
+#' 
+#' @param input An actel results object, or a preload object
+#' @param arrays Optional: A subset of arrays to be plotted
+#' @param show.stations Logical: Should the live times of each station be shown under the array bars?
+#' @param array.size The size of the array bars (defaults to 2)
+#' @param station.size The size of the station bars (defaults to 1)
+#' @param show.caps Logical: Should cap lines be shown at the end of each live period?
+#' @param cap.prop The relative size of the caps, as compared to the respective bars (defaults to 2).
+#' @param title An optional title for the plot.
+#' @param xlab,ylab Optional axis names for the plot.
+#' @param col An optional colour scheme for the array bars. If left empty, default colours will be added.
+#'  Note: Station bars are 40% lighter than the array bars.
+#' 
+#' @return A ggplot object.
+#'
+#' @examples
+#' # Using the example results that come with actel
+#' plotLive(example.results)
+#'
+#' # Because plotSensors returns a ggplot object, you can store
+#' # it and edit it manually, e.g.:
+#' library(ggplot2)
+#' p <- plotLive(example.results)
+#' p <- p + xlab("changed the x axis label a posteriori")
+#' p
+#'
+#' # You can also save the plot using ggsave!
+#'
+#' @export
+#'
+#' 
+plotLive <- function(input, arrays, show.stations = FALSE, array.size = 2, station.size = 1, 
+                     show.caps = TRUE, cap.prop = 2, title = "", xlab = "", ylab = "", col) {
+  value <- NULL
+  Y <- NULL
+  line <- NULL
+  Section <- NULL
+
+  cbPalette <- c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7")
+  names(cbPalette) <- c("Orange", "Blue", "Green", "Yellow", "Darkblue", "Darkorange", "Pink")
+
+  if (!inherits(input, "list"))
+    stop("Could not recognise the input as an actel results or preload object.", call. = FALSE)
+
+  if (is.null(input$spatial) | is.null(input$arrays) | is.null(input$deployments))
+    stop("Could not recognise the input as an actel results or preload object.", call. = FALSE)
+
+  if (!missing(arrays) && any(is.na(match(arrays, unlist(input$spatial$station.order)))))
+    stop("'arrays' was set but not all contents match array names in the study area.", call. = FALSE)
+
+  spatial <- input$spatial
+  deployments <- input$deployments
+  study.arrays <- input$arrays
+
+  # prepare input
+  if (show.stations) {
+    xdep <- do.call(rbind, deployments)
+    xdep$line <- 1:nrow(xdep)
+    pd <- reshape2::melt(xdep[, c("Array", "line", "Standard.name", "Start", "Stop")], id.vars = c("Array", "line", "Standard.name"))
+    colnames(pd)[colnames(pd) == "Standard.name"] <- "Y"
+
+    aux <- lapply(study.arrays, function(x) x$live)
+    aux <- data.table::rbindlist(aux, idcol = "Array")
+    aux$Y <- aux$Array
+    aux$line <- (nrow(xdep) + 1):(nrow(xdep) + nrow(aux))
+    to.add <- reshape2::melt(aux, id.vars = c("Array", "line", "Y"))
+
+    pd <- rbind(pd, to.add)
+  } else {
+    aux <- lapply(study.arrays, function(x) x$live)
+    aux <- data.table::rbindlist(aux, idcol = "Array")
+    aux$Y <- aux$Array
+    aux$line <- 1:nrow(aux)
+    pd <- reshape2::melt(aux, id.vars = c("Array", "line", "Y"))
+  }
+
+  # subset if needed
+  if (!missing(arrays)) {
+    pd <- pd[pd$Array %in% arrays, ]
+  }
+
+  # find y.order
+  if (show.stations) {
+    x <- split(spatial$stations$Standard.name, spatial$stations$Array)
+    x <- x[unlist(spatial$array.order)]
+
+    y.order <- unlist(lapply(names(x), function(i) {
+      c(i, sort(x[[i]]))
+    }))
+
+    y.order <- rev(y.order)
+  } else {
+    y.order <- rev(unlist(spatial$array.order))
+  }
+
+  pd$Y <- factor(pd$Y, levels = y.order)
+  pd$Y <- droplevels(pd$Y)
+
+  # assign sections
+  link <- lapply(1:length(spatial$array.order), function(i) {
+    output <- rep(NA_real_, nrow(pd))
+    output[pd$Array %in% spatial$array.order[[i]]] <- i
+    return(output)
+  })
+  link <- combine(link)
+  pd$Section <- factor(names(spatial$array.order)[link], levels = names(spatial$array.order))
+
+  # choose colours
+  if (missing(col)) {
+    if (length(names(spatial$array.order)) <= 7) {
+      col <- as.vector(cbPalette)[1:length(names(spatial$array.order))]
+    } else {
+      col <- gg_colour_hue(length(names(spatial$array.order)))
+    }
+  } else {
+    if (length(col) < length(names(spatial$array.order))) {
+      warning("Not enough colours supplied in 'col' (", length(col)," supplied and ", length(names(spatial$array.order)), " needed). Reusing colours.", immediate. = TRUE, call. = FALSE)
+      col <- rep(col, length.out = length(names(spatial$array.order)))
+    }
+  }
+  st.col <- darken(col, 0.6)
+
+  # assing colours
+  link <- match(pd$Section, names(spatial$array.order))
+  pd$col <- sapply(1:nrow(pd), function(i) {
+    if (pd$Array[i] == pd$Y[i])
+      col[link[i]]
+    else
+      st.col[link[i]]
+  })
+
+  # assign sizes
+  pd$Size <- ifelse(pd$Array == pd$Y, array.size, station.size)
+
+  # sort pd
+  pd <- pd[order(pd$Y), ]
+
+  p <- ggplot2::ggplot(data = pd, ggplot2::aes(x = value, y = Y, group = line))
+  # place holder just to set the Y in order
+  p <- p + ggplot2::geom_path()
+  
+  a <- pd[pd$Array == pd$Y, ]
+  p <- p + ggplot2::geom_path(data = a, ggplot2::aes(col = Section), size = a$Size)
+  
+  if (show.stations) {
+    s <- pd[pd$Array != pd$Y, ]
+    p <- p + ggplot2::geom_path(data = s, col = s$col, size = s$Size)
+  }
+
+  if (show.caps)
+    p <- p + ggplot2::geom_point(shape = "|", size = pd$Size*cap.prop, colour = pd$col)
+
+  p <- p + ggplot2::theme_bw()
+  p <- p + ggplot2::labs(title = title, x = xlab, y = ylab)
+  p <- p + ggplot2::scale_color_manual(values = col)
+  p <- p + ggplot2::guides(colour = ggplot2::guide_legend(override.aes = list(size = array.size)))
+  p
+}
+
 #' Plot sensor data for a single tag
 #'
 #' The output of plotSensors is a ggplot object, which means you can then use it in combination